🐛 fix(post write): rename selectTagOrganism component
