+++ src/screens/RecommendScreen/PostWriteScreen/PostWriteMainScreen.tsx
@@ -2,18 +2,18 @@ import {NativeStackScreenProps} from '@react-navigation/native-stack';
 import React, {useLayoutEffect} from 'react';
 import {SafeAreaView} from 'react-native';
 
-import {RecommendParamList} from '..';
+import {PostWriteParamList} from '.';
 
 import AddPhotoOrganism from 'src/components/PostWrite/AddPhotoOrganism';
 import InputContentsOrganism from 'src/components/PostWrite/InputContentsOrganism';
 import InputCustomTagOrganism from 'src/components/PostWrite/InputCustomTagOrganism';
-import SelectTagOrganism from 'src/components/PostWrite/SelectTagOrganism';
+import MoveSelectTagOrganism from 'src/components/PostWrite/MoveSelectTagOrganism';
 import SubmitOrganism from 'src/components/PostWrite/SubmitOrganism';
 import DismissKeyboardView from 'src/components/utils/DismissKeyboardScrollView';
 import LeftBackHeader from 'src/components/utils/Header/LeftBackHeader';
 import {heightPercentage} from 'src/styles/ScreenResponse';
 
-export type DetailScreenProps = NativeStackScreenProps<RecommendParamList, 'PostWrite'>;
+export type DetailScreenProps = NativeStackScreenProps<PostWriteParamList, 'PostWriteMain'>;
 
 const PostWriteMainScreen = ({navigation}: DetailScreenProps) => {
   useLayoutEffect(() => {
@@ -28,7 +28,7 @@ const PostWriteMainScreen = ({navigation}: DetailScreenProps) => {
       <DismissKeyboardView>
         <AddPhotoOrganism />
         <InputContentsOrganism />
-        <SelectTagOrganism />
+        <MoveSelectTagOrganism />
         <InputCustomTagOrganism />
         <SubmitOrganism />
       </DismissKeyboardView>
